requirements.txt
pytest
pytest-bdd
pytest-xdist
pytest-html
allure-pytest
playwright
requests
httpx
sqlalchemy
psycopg2-binary
python-dotenv
